@@ -3,200 +3,180 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<parent>
-		<groupId>org.springframework.boot</groupId>
-		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.0.2.RELEASE</version>
-		<relativePath /> <!-- lookup parent from repository -->
-	</parent>
-	
 	<groupId>io.mosip.mock.mds</groupId>
 	<artifactId>mock-mds</artifactId>
-	<version>1.2.1-SNAPSHOT</version>
+	<version>1.2.1-java21-SNAPSHOT</version>
 	<name>mock-mds</name>
 	<description>A mock project for biometric provider</description>
+	<url>https://github.com/mosip/mosip-mock-services</url>
 
-	<licenses>
-		<license>
-			<name>MPL 2.0</name>
-			<url>https://www.mozilla.org/en-US/MPL/2.0/</url>
-		</license>
-	</licenses>
-	<scm>
-		<connection>scm:git:git://github.com/mosip/mosip-ref-impl.git</connection>
-		<developerConnection>scm:git:ssh://github.com:mosip/mosip-ref-impl.git</developerConnection>
-		<url>https://github.com/mosip/mosip-ref-impl</url>
-		<tag>HEAD</tag>
-	</scm>
+	<properties>
+		<doclint>none</doclint>
 
-	<developers>
-		<developer>
-			<name>Mosip</name>
-			<email>mosip.emailnotifier@gmail.com</email>
-			<organization>io.mosip</organization>
-			<organizationUrl>https://github.com/mosip/mosip-mock-services</organizationUrl>
-		</developer>
-	</developers>
+		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+		<java.version>21</java.version>
+
+		<!-- maven -->
+		<maven.compiler.source>${java.version}</maven.compiler.source>
+		<maven.compiler.target>${java.version}</maven.compiler.target>
+		<maven.compiler.version>3.11.0</maven.compiler.version>
+		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
+		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
+		<maven.javadoc.plugin.version>3.2.0</maven.javadoc.plugin.version>
+		<maven.shade.plugin.version>2.3</maven.shade.plugin.version>
+		<jacoco.maven.plugin.version>0.8.11</jacoco.maven.plugin.version>
+		<maven.deploy.plugin.version>3.1.1</maven.deploy.plugin.version>
+		<maven.source.plugin.version>3.3.1</maven.source.plugin.version>
+		<maven.javadoc.plugin.version>3.2.0</maven.javadoc.plugin.version>
+		<maven.gpg.plugin.version>3.2.3</maven.gpg.plugin.version>
+		<maven.jar.plugin.version>3.0.2</maven.jar.plugin.version>
+		<maven.war.plugin.version>3.1.0</maven.war.plugin.version>
+		<maven.dependency.plugin.version>3.1.2</maven.dependency.plugin.version>
+		<maven.resources.plugin.version>3.0.1</maven.resources.plugin.version>
+		<maven.surefire.plugin.version>2.22.0</maven.surefire.plugin.version>
+		<nexus.staging.maven.plugin.version>1.6.7</nexus.staging.maven.plugin.version>
+		<spring.boot.maven.plugin.version>3.2.5</spring.boot.maven.plugin.version>
+
+		<!-- git -->
+		<git.commit.id.plugin.version>3.0.1</git.commit.id.plugin.version>
+
+		<!-- Others -->
+		<jai.imageio.core.version>1.3.0</jai.imageio.core.version>
+		<org.apache.httpmime.version>4.5.7</org.apache.httpmime.version>
+		<com.squareup.okhttp.version>2.7.5</com.squareup.okhttp.version>
+		<org.codehaus.jackson.mapper.asl.version>1.9.13</org.codehaus.jackson.mapper.asl.version>
+		<org.bouncycastle.bcprov.jdk18on.version>1.78.1</org.bouncycastle.bcprov.jdk18on.version>
+		<javax.xml.crypto.jsr105.api.version>1.0.1</javax.xml.crypto.jsr105.api.version>
+		<org.bitbucket.jose4j.version>0.7.1</org.bitbucket.jose4j.version>
+
+		<!-- Mosip -->
+		<kernel.bom.version>1.2.1-java21-SNAPSHOT</kernel.bom.version>
+		<kernel.core.version>1.2.1-java21-SNAPSHOT</kernel.core.version>
+		<kernel.crypto.jce.version>1.0.6</kernel.crypto.jce.version>
+
+		<mosip.biometrics.util.version>1.2.1-java21-SNAPSHOT</mosip.biometrics.util.version>
+	</properties>
+
+	<dependencyManagement>
+		<dependencies>
+			<dependency>
+				<groupId>io.mosip.kernel</groupId>
+				<artifactId>kernel-bom</artifactId>
+				<version>${kernel.bom.version}</version>
+				<type>pom</type>
+				<scope>import</scope>
+			</dependency>
+		</dependencies>
+	</dependencyManagement>
+
 	<dependencies>
 		<dependency>
-		    <groupId>com.github.jai-imageio</groupId>
-		    <artifactId>jai-imageio-core</artifactId>
-		    <version>1.3.0</version>
-		</dependency>		
-    	<dependency>
+			<groupId>com.github.jai-imageio</groupId>
+			<artifactId>jai-imageio-core</artifactId>
+			<version>${jai.imageio.core.version}</version>
+		</dependency>
+		<dependency>
 			<groupId>io.mosip.biometric.util</groupId>
-		  	<artifactId>biometrics-util</artifactId>
-		  	<version>1.2.1-SNAPSHOT</version>
+			<artifactId>biometrics-util</artifactId>
+			<version>${mosip.biometrics.util.version}</version>
 		</dependency>
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
 			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
-
-		 <dependency>
-          <groupId>org.projectlombok</groupId>
-          <artifactId>lombok</artifactId>
-          <version>RELEASE</version>
-          <scope>compile</scope>
-      	</dependency>
-
+		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+			<scope>provided</scope>
+		</dependency>
 		<dependency>
 			<groupId>com.squareup.okhttp</groupId>
 			<artifactId>okhttp</artifactId>
-			<version>2.7.5</version>
+			<version>${com.squareup.okhttp.version}</version>
 		</dependency>
-
 		<dependency>
 			<groupId>org.apache.httpcomponents</groupId>
 			<artifactId>httpclient</artifactId>
-			<version>4.5.7</version>
 		</dependency>
-
 		<dependency>
 			<groupId>org.apache.httpcomponents</groupId>
 			<artifactId>httpmime</artifactId>
-			<version>4.5.7</version>
+			<version>${org.apache.httpmime.version}</version>
 		</dependency>
-
-		<!-- https://mvnrepository.com/artifact/org.codehaus.jackson/jackson-mapper-asl -->
 		<dependency>
 			<groupId>org.codehaus.jackson</groupId>
 			<artifactId>jackson-mapper-asl</artifactId>
-			<version>1.9.13</version>
+			<version>${org.codehaus.jackson.mapper.asl.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>javax.xml.bind</groupId>
+			<artifactId>jaxb-api</artifactId>
+		</dependency>
+		<dependency>
+		    <groupId>jakarta.servlet</groupId>
+		    <artifactId>jakarta.servlet-api</artifactId>
+		    <scope>provided</scope>
+		</dependency>
+		<dependency>
+			<groupId>io.mosip.kernel</groupId>
+			<artifactId>kernel-crypto-jce</artifactId>
+			<version>${kernel.crypto.jce.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.apache.commons</groupId>
+			<artifactId>commons-lang3</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.bouncycastle</groupId>
+			<artifactId>bcprov-jdk18on</artifactId>
+			<version>${org.bouncycastle.bcprov.jdk18on.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>javax.xml.crypto</groupId>
+			<artifactId>jsr105-api</artifactId>
+			<version>${javax.xml.crypto.jsr105.api.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>io.mosip.kernel</groupId>
+			<artifactId>kernel-core</artifactId>
+			<version>${kernel.core.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>org.bitbucket.b_c</groupId>
+			<artifactId>jose4j</artifactId>
+			<version>${org.bitbucket.jose4j.version}</version>
 		</dependency>
 
-
-		<!-- https://mvnrepository.com/artifact/com.squareup.okhttp/okhttp -->
-			<dependency>
-				<groupId>com.squareup.okhttp</groupId>
-				<artifactId>okhttp</artifactId>
-				<version>2.7.5</version>
-			</dependency>
-
-			<!-- https://mvnrepository.com/artifact/javax.xml.bind/jaxb-api -->
-			<dependency>
-			    <groupId>javax.xml.bind</groupId>
-			    <artifactId>jaxb-api</artifactId>
-			    <version>2.3.1</version>
-			</dependency>
-		
-		   <dependency>
-				<groupId>io.mosip.kernel</groupId>
-				<artifactId>kernel-crypto-jce</artifactId>
-				<version>1.0.6</version>
-			</dependency>
-			<!-- https://mvnrepository.com/artifact/org.apache.commons/commons-lang3 -->
-			<dependency>
-				<groupId>org.apache.commons</groupId>
-				<artifactId>commons-lang3</artifactId>
-				<version>3.10</version>
-			</dependency>
-			
-			<dependency>
-			    <groupId>org.bouncycastle</groupId>
-			    <artifactId>bcprov-jdk16</artifactId>
-			    <version>1.46</version>
-			</dependency>
-			
-			<!-- https://mvnrepository.com/artifact/javax.xml.crypto/jsr105-api -->
-			<dependency>
-			    <groupId>javax.xml.crypto</groupId>
-			    <artifactId>jsr105-api</artifactId>
-			    <version>1.0.1</version>
-			</dependency>
-			
-			<dependency>
-				<groupId>io.mosip.kernel</groupId>
-				<artifactId>kernel-core</artifactId>
-				<version>1.2.1-SNAPSHOT</version>
-			</dependency>
-
-			<!-- https://mvnrepository.com/artifact/org.bitbucket.b_c/jose4j -->
-			<dependency>
-			    <groupId>org.bitbucket.b_c</groupId>
-			    <artifactId>jose4j</artifactId>
-			    <version>0.7.1</version>
-			</dependency>
-
 	</dependencies>
-	<repositories>
+
+	<distributionManagement>
+		<snapshotRepository>
+			<id>ossrh</id>
+			<url>https://oss.sonatype.org/content/repositories/snapshots</url>
+		</snapshotRepository>
 		<repository>
-	         <id>snapshots-repo</id>
-	         <url>https://oss.sonatype.org/content/repositories/snapshots</url>
-	         <releases><enabled>false</enabled></releases>
-	         <snapshots><enabled>true</enabled></snapshots>
-       	</repository>
-
-        <repository>
-	         <id>releases-repo</id>
-	         <url>https://oss.sonatype.org/service/local/staging/deploy/maven2</url>
-	         <releases><enabled>false</enabled></releases>
-	         <snapshots><enabled>false</enabled></snapshots>
-       	</repository>
-
-        <repository>
-	        <id>danubetech-maven-public</id>
-	        <url>https://repo.danubetech.com/repository/maven-public/</url>
-      	</repository>
-	</repositories>
+			<id>ossrh</id>
+			<url>https://oss.sonatype.org/service/local/staging/deploy/maven2/</url>
+		</repository>
+	</distributionManagement>
 
 	<build>
 		<plugins>
-			<!-- <plugin> <groupId>org.springframework.boot</groupId> <artifactId>spring-boot-maven-plugin</artifactId> 
-				<version>2.2.6.RELEASE</version> <executions> <execution> <goals> <goal>repackage</goal> 
-				</goals> </execution> </executions> <configuration> <release>11</release> 
-				<fork>true</fork> <meminitial>1024m</meminitial> <maxmem>2024m</maxmem> <archive> 
-				<manifest> <mainClass>org.biometric.provider.ProviderApplication</mainClass> 
-				</manifest> </archive> </configuration> </plugin> -->
 			<plugin>
-				<groupId>pl.project13.maven</groupId>
-				<artifactId>git-commit-id-plugin</artifactId>
-				<version>3.0.1</version>
-				<executions>
-					<execution>
-						<id>get-the-git-infos</id>
-						<goals>
-							<goal>revision</goal>
-						</goals>
-						<phase>validate</phase>
-					</execution>
-				</executions>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-compiler-plugin</artifactId>
+				<version>${maven.compiler.version}</version>
 				<configuration>
-					<generateGitPropertiesFile>true</generateGitPropertiesFile>
-					<generateGitPropertiesFilename>${project.build.outputDirectory}/git.properties</generateGitPropertiesFilename>
-					<includeOnlyProperties>
-						<includeOnlyProperty>^git.build.(time|version)$</includeOnlyProperty>
-						<includeOnlyProperty>^git.commit.id.(abbrev|full)$</includeOnlyProperty>
-					</includeOnlyProperties>
-					<commitIdGenerationMode>full</commitIdGenerationMode>
-					<dotGitDirectory>${project.basedir}/.git</dotGitDirectory>
-					<!-- <format>json</format> -->
+					<release>${java.version}</release>
+					<compilerArgs>
+						<arg>--enable-preview</arg>
+					</compilerArgs>
 				</configuration>
 			</plugin>
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
 				<artifactId>maven-jar-plugin</artifactId>
-				<version>2.4</version>
+				<version>${maven.jar.plugin.version}</version>
 				<configuration>
 					<archive>
 						<manifest>
