@@ -15,16 +15,16 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
-import javax.jms.BytesMessage;
-import javax.jms.Connection;
-import javax.jms.Destination;
-import javax.jms.JMSException;
-import javax.jms.Message;
-import javax.jms.MessageConsumer;
-import javax.jms.MessageListener;
-import javax.jms.MessageProducer;
-import javax.jms.Session;
-import javax.jms.TextMessage;
+import jakarta.jms.BytesMessage;
+import jakarta.jms.Connection;
+import jakarta.jms.Destination;
+import jakarta.jms.JMSException;
+import jakarta.jms.Message;
+import jakarta.jms.MessageConsumer;
+import jakarta.jms.MessageListener;
+import jakarta.jms.MessageProducer;
+import jakarta.jms.Session;
+import jakarta.jms.TextMessage;
 
 import org.apache.activemq.ActiveMQConnection;
 import org.apache.activemq.ActiveMQConnectionFactory;
@@ -130,7 +130,7 @@ public class Listener {
 
 	public String outBoundQueue;
 
-	public void consumeLogic(javax.jms.Message message, String abismiddlewareaddress) {
+	public void consumeLogic(jakarta.jms.Message message, String abismiddlewareaddress) {
 		ResponseEntity<Object> obj = null;
 		Map map = null;
 		Integer textType = 0;
@@ -473,7 +473,7 @@ public class Listener {
 					QueueListener listener = new QueueListener() {
 
 						@Override
-						public void setListener(javax.jms.Message message) {
+						public void setListener(jakarta.jms.Message message) {
 							consumeLogic(message, outBoundAddress);
 						}
 					};
