@@ -1,6 +1,6 @@
 package io.mosip.proxy.abis;
 
-import javax.jms.Message;
+import jakarta.jms.Message;
 
 public abstract class QueueListener {	
 	public abstract void setListener(Message message);
